Реализовал адаптив для секций work, footer
