+++ src/main.js
@@ -5,3 +5,5 @@ if (process.env.NODE_ENV === "development") {
 
 import "./scripts/skills";
 import "./scripts/hamburger-menu";
+import "./scripts/parallax";
+import "./scripts/parallax-reviews";
