+++ src/mixins.pug
@@ -3,10 +3,6 @@ mixin icon(name, className)
   svg(class=className viewBox=icon.viewBox preserveAspectRatio="none")
     use(xlink:href=icon.url)
 
-//- mixin image(path, className, description, attrRole)
-//-   img(src=require(`images/${path}`) class=className)
-
-
 mixin image(path, className, description, attrRole)
   if role
     img(src=require(`images/${path}`) class=className role= attrRole)
