+++ src/scripts/hamburger-menu.js
@@ -12,6 +12,3 @@ hamburgerButton.addEventListener("click", function () {
     promo.classList.toggle("promo-hamburger--active")
     body.classList.toggle("body-noscroll")
 })
-
-
-//
