+++ src/styles/blocks/nav.pcss
@@ -31,3 +31,43 @@
 .hamburger-nav__item:hover {
   color: $purple;
 }
+
+/* footer nav */
+
+/* .nav--footer {
+  @include tablets {
+} */
+
+.nav--footer__list {
+  display: flex;
+  @include desktop {
+    justify-content: space-around;
+  }
+  @include tablets {
+    justify-content: end;
+    flex-wrap: wrap;
+  }
+}
+
+.nav--footer__item {
+  margin-right: 8%;
+  @include desktop {
+    margin-right: 7%;
+  }
+  @include tablets {
+    margin-right: 10%;
+    margin-bottom: 15px;
+  }
+
+  &:last-child {
+    margin-right: 0;
+  }
+}
+
+.nav--footer__link {
+  text-decoration: none;
+}
+
+.nav--footer__link:hover {
+  color: $purple;
+}
