+++ src/styles/blocks/parallax-reviews.pcss
@@ -1,55 +1,55 @@
 .parallax-reviews {
   position: relative;
   overflow: hidden;
+  @include tablets {
+    background: url(/src/images/bg/Budda.jpg) center center;
+    .parallax-reviews__layer {
+      display: none;
+    }
+  }
 }
 
-.parallax__image-reviews {
-  position: absolute;
-}
-
-.parallax__layer-reviews {
+.parallax-reviews__layer {
   position: absolute;
   top: 0;
   left: 0;
-  width: 100%;
+  width: 120%;
   height: 100%;
-  &:nth-child(1) {
-    .parallax__image-reviews {
-      object-fit: cover;
-      width: 100%;
-      height: 100%;
-    }
-  }
-  &:nth-child(2) {
-    .parallax__image-reviews {
-      top: 700px;
-    }
-  }
-  &:nth-child(3) {
-    .parallax__image-reviews {
-      bottom: 170px;
-      left: 60%;
-    }
-  }
-  &:nth-child(4) {
-    .parallax__image-reviews {
-      bottom: -800px;
-      left: 0;
-      height: 100%;
-      width: 100%;
-      object-fit: cover;
-    }
-  }
-  &:nth-child(5) {
-    .parallax__image-reviews {
-      bottom: -635px;
-      left: 660px;
-      width: 100%;
-      object-fit: cover;
-    }
-  }
+  left: -20%;
 }
 
-.parallax__image-reviews {
+.parallax__image-reviews-sky {
+  object-fit: cover;
+  width: 120%;
+  height: 100%;
   position: absolute;
+}
+
+.parallax__image-reviews-mountain {
+  top: 500px;
+  position: absolute;
+  width: 130%;
+}
+
+.parallax__image-reviews-buddha {
+  bottom: 170px;
+  left: 60%;
+  position: absolute;
+}
+
+.parallax__image-reviews-white {
+  bottom: -800px;
+  left: 0;
+  height: 100%;
+  object-fit: cover;
+  position: absolute;
+  width: 130%;
+}
+
+.parallax__image-reviews-clouds1 {
+  bottom: -635px;
+  left: 660px;
+  object-fit: cover;
+  position: absolute;
+  width: 130%;
 }
