+++ src/styles/blocks/slider-control-work.pcss
@@ -3,6 +3,9 @@
   position: absolute;
   right: 0;
   background-color: black;
+  @include phones {
+    bottom: 0;
+  }
 }
 
 .slider__btn--prev,
