+++ src/styles/blocks/slider-reviews.pcss
@@ -4,7 +4,6 @@
     overflow: hidden;
   }
   @include m-phones {
-    /* width: 380px; */
     overflow: hidden;
   }
 }
@@ -24,6 +23,7 @@
   max-width: 300px;
   @include phones {
     min-width: 340px;
+    width: 100vw;
   }
   @include m-phones {
     min-width: 260px;
