+++ src/styles/blocks/social.pcss
@@ -33,14 +33,15 @@
   border-radius: 50%;
   margin-right: 25px;
   position: relative;
-  opacity: 0.1;
+  opacity: 0.2;
+  cursor: pointer;
   &:last-child {
     margin-right: 0;
   }
   &:hover {
-    opacity: 0.8;
+    opacity: 1;
     .social__icon--footer {
-      opacity: 0.8;
+      opacity: 1;
     }
   }
 }
@@ -53,7 +54,7 @@
   width: 20px;
   height: 20px;
   fill: $white;
-  opacity: 0.5;
+  /* opacity: 0.5; */
   &:hover {
     opacity: 1;
   }
