+++ src/styles/blocks/work.pcss
@@ -6,10 +6,8 @@
 
 .work__content {
   display: flex;
-  @include desktop {
+  @include tablets {
     flex-direction: column;
-    align-items: center;
-    position: relative;
   }
 }
