+++ src/styles/layout/base.pcss
@@ -80,6 +80,14 @@ button {
   &__work {
     margin-bottom: 50px;
   }
+  &__skills {
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 48px;
+    @include tablets {
+      text-align: center;
+    }
+  }
 }
 
 .text {
