+++ src/templates/footer.pug
@@ -1,18 +1,35 @@
-footer.footer
+.footer#footer
     .container.footer__container
-        .footer__base
-            .footer__year &copy 2020
-            .footer__info-user Юрий <br> Кочергин
-                .footer__info-text.text Личный сайт <br> веб-разработчика
-        .footer__bar
+        .footer__info 
+            span.footer__copy &copy2020
+            .footer__user 
+                .footer__user-name Юрий Кочергин
+                .footer__user-site Личный сайт веб-разработчика
+        .footer__content  
             .footer__nav
-                +nav("nav", ["Главная", "Обо мне", "Навыки", "Работы", "Отзывы", "Связаться"], ["promo", "about", "skill", "work", "reviews", "contact"])
-            .footer__info-desc
-                .footer__info-base
-                    .footer__text.text Этот сайт я сделал в рамках обучения в Школе онлайн образования Loftschool
+                include ./nav.pug
+                +nav("nav--footer", ["Главная", "Обо мне", "Навыки", "Работы", "Отзывы", "Связаться"], ["promo", "about", "skill", "work", "reviews", "contact"])
+            .footer__desc
+                .footer__box
+                    .footer__text
+                        p Этот сайт я сделал в рамках обучения в Школе онлайн образования Loftschool
                     .footer__social
                         include ./social.pug
-                        +social("footer")  
-                .footer__info-bar
-                    .footer__text.text Всегда пишите код так, будто сопровождать его будет склонный к насилию психопат, который знает где вы живете
-                    .footer__text-autor.text Martin Golding
+                        +social("footer") 
+                .footer__box
+                    .footer__text
+                        p Всегда пишите код так, будто сопровождать его будет склонный к насилию психопат, который знает где вы живете
+                    .footer__autor
+                        span Martin Golding
+
+                
+
+
+
+
+
+
+
+
+
+
