+++ src/templates/skills.pug
@@ -1,23 +1,67 @@
 .skills#skills
-    h4.skill__title.title Front-end
-    ul.skill__list
-        li.skill__item
-            .skill__item-name HTML5
-        li.skill__item
-            .skill__item-name CSS4
-        li.skill__item
-            .skill__item-name Javascript
-        li.skill__item
-            .skill__item-name.skill__item-name Jquery <br>и Vue.js
-    h4.skill__title.title Workflow
-    ul.skill__list
-        li.skill__item
-            .skill__item-name GIT
-        li.skill__item
-            .skill__item-name Terminal
-        li.skill__item
-             .skill__item-name Gulp
-        li.skill__item
-            .skill__item-name Webpack
+    h2.title.title__skills Front-End
+    ul.skills__list
+        li.skills__item 
+            .skills__box
+                .skills__box--title HTML5
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title CSS
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title Javascript
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title query и Vue.js
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+    h2.title.title__skills Workflow
+    ul.skills__list
+        li.skills__item
+            .skills__box
+                .skills__box--title Git
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title Terminal
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title Gulp
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
+        li.skills__item
+            .skills__box
+                .skills__box--title Webpack
+                svg(viewBox="0 0 200 200").skills__box--svg-shape
+                    circle(cx="100" cy="100" r="80").skills__box--svg-circle--position-under
+                    circle(cx="100" cy="100" r="80" style=`animation-delay: -50s`).skills__box--svg-circle--position-above
 
-            
+
+
+
+svg(style="width:0; height:0; position: absolute;" aria-hidden="true")
+    linearGradient(id="gradient" x1="0%" y1="0%" x2="100%" y2="0")
+        stop(offset="0%" stop-color='#4a00ed')
+        stop(offset="100%" stop-color='#9300e8') 
+    
+    
+    
+    
+    
+
